Move BFT classes out of Tempo
