+++ radixdlt/src/main/java/org/radix/Radix.java
@@ -17,8 +17,8 @@
 
 package org.radix;
 
-import com.radixdlt.consensus.tempo.DumbPacemaker;
-import com.radixdlt.consensus.tempo.MemPool;
+import com.radixdlt.consensus.DumbPacemaker;
+import com.radixdlt.consensus.MemPool;
 import com.radixdlt.middleware2.converters.AtomToBinaryConverter;
 import com.radixdlt.serialization.Serialization;
 import com.radixdlt.serialization.SerializationException;
